Add files via upload
Skilled Data Scientist with 5 years of experience developing and implementing analytic models to improve business outcomes. Successfully development of predictive models, resulting in a 20% increase in revenue and a 15% increase in customer retention. Implemented natural language processing models to enhance customer service interactions, resulting in a 15% decrease in customer complaints.